Renamed DynamicUrlParamActionInvoker to FeatherActionInvoker. Exposed a method for overriding the behaviour when an exception occurs while executing a controller
