+++ Telerik.Sitefinity.Frontend/Telerik.Sitefinity.Frontend.csproj
@@ -352,7 +352,7 @@
     </Compile>
     <Compile Include="Mvc\Infrastructure\Routing\DefaultUrlParamsMapper.cs" />
     <Compile Include="Mvc\Infrastructure\Routing\DetailActionParamsMapper.cs" />
-    <Compile Include="Mvc\Infrastructure\Routing\DynamicUrlParamActionInvoker.cs" />
+    <Compile Include="Mvc\Infrastructure\Routing\FeatherActionInvoker.cs" />
     <Compile Include="Mvc\Infrastructure\Routing\IRouteParamResolver.cs" />
     <Compile Include="Mvc\Infrastructure\Routing\IUrlParamsMapper.cs" />
     <Compile Include="Mvc\Infrastructure\Routing\RouteParamResolverBase.cs" />
@@ -760,12 +760,14 @@
   <Import Project="$(SolutionDir)\StyleCop\StyleCop.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties OpenAccess_EnhancementOutputLevel="1" OpenAccess_UpdateDatabase="False" OpenAccess_Enhancing="False" OpenAccess_ConnectionId="DatabaseConnection1" OpenAccess_ConfigFile="App.config" />
+      <UserProperties OpenAccess_ConfigFile="App.config" OpenAccess_ConnectionId="DatabaseConnection1" OpenAccess_Enhancing="False" OpenAccess_UpdateDatabase="False" OpenAccess_EnhancementOutputLevel="1" />
     </VisualStudio>
   </ProjectExtensions>
   <PropertyGroup>
-    <PostBuildEvent></PostBuildEvent>
-    <PreBuildEvent></PreBuildEvent>
+    <PostBuildEvent>
+    </PostBuildEvent>
+    <PreBuildEvent>
+    </PreBuildEvent>
   </PropertyGroup>
   <!-- Telerik OpenAccess Enhancement -->
   <!-- Fluent API Mapping -->
